+++ src/main/java/io/elastic/jdbc/actions/UpsertRowByPrimaryKey.java
@@ -11,6 +11,7 @@ import java.sql.Connection;
 import java.sql.DatabaseMetaData;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.List;
 import javax.json.Json;
 import javax.json.JsonObject;
 import org.slf4j.Logger;
@@ -99,6 +100,14 @@ public class UpsertRowByPrimaryKey implements Module {
         }
       }
     } catch (SQLException e) {
+      if (Utils.reboundIsEnabled(configuration)) {
+        List<String> states = Utils.reboundDbState.get(dbEngine);
+        if (states.contains(e.getSQLState())) {
+          LOGGER.warn("Starting rebound max iter: {}, rebound ttl: {}. Reason: {}", System.getenv("ELASTICIO_REBOUND_LIMIT"), System.getenv("ELASTICIO_REBOUND_INITIAL_EXPIRATION"), e.getMessage());
+          parameters.getEventEmitter().emitRebound(e);
+          return;
+        }
+      }
       throw new RuntimeException(e);
     }
   }